+++ test/ai/assistant.test.ts
@@ -1,8 +1,8 @@
-import AinftJs from '../src/ainft';
-import { AssistantCreateParams, AssistantUpdateParams } from '../src/types';
-import { privateKey, objectId, tokenId, assistantId } from './test_data';
+import AinftJs from '../../src/ainft';
+import { AssistantCreateParams, AssistantUpdateParams } from '../../src/types';
+import { privateKey, objectId, tokenId, assistantId } from '../test_data';
 
-jest.mock('../src/common/util', () => {
+jest.mock('../src/utils/util', () => {
   const mockRequest = jest.fn((jobType, body) => {
     switch (jobType) {
       case 'create_assistant':
@@ -27,7 +27,7 @@ jest.mock('../src/common/util', () => {
         return null;
     }
   });
-  const util = jest.requireActual('../src/common/util');
+  const util = jest.requireActual('../src/utils/util');
   return {
     ...util,
     validateAssistant: jest.fn().mockResolvedValue(undefined),
